+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2 (December 15, 2019)
+
+* Add copy to clipboard button
+* Add change visibility button
+
 ## 1.0.1 (December 15, 2019)
 
 Initial release